Remove dependence on statusProvider
